Rename back to browser
k6 depends on the import to be browser and not module.
