+++ browser/registry_test.go
@@ -1,4 +1,4 @@
-package module
+package browser
 
 import (
 	"sync"
