+++ register.go
@@ -7,7 +7,7 @@ import (
 	_ "net/http/pprof" //nolint:gosec
 	"os"
 
-	"github.com/grafana/xk6-browser/module"
+	"github.com/grafana/xk6-browser/browser"
 
 	k6modules "go.k6.io/k6/js/modules"
 )
@@ -24,5 +24,5 @@ func init() {
 }
 
 func init() {
-	k6modules.Register("k6/x/browser", module.New())
+	k6modules.Register("k6/x/browser", browser.New())
 }
