+++ tests/browser_type_test.go
@@ -5,9 +5,9 @@ import (
 
 	"github.com/stretchr/testify/require"
 
+	"github.com/grafana/xk6-browser/browser"
 	"github.com/grafana/xk6-browser/chromium"
 	"github.com/grafana/xk6-browser/k6ext/k6test"
-	"github.com/grafana/xk6-browser/module"
 )
 
 func TestBrowserTypeConnect(t *testing.T) {
@@ -35,9 +35,9 @@ func TestBrowserTypeLaunchToConnect(t *testing.T) {
 	// to take mapping layer into account, instead of calling
 	// BrowserType.Launch method directly
 	rt := vu.Runtime()
-	root := module.New()
+	root := browser.New()
 	mod := root.NewModuleInstance(vu)
-	jsMod, ok := mod.Exports().Default.(*module.JSModule)
+	jsMod, ok := mod.Exports().Default.(*browser.JSModule)
 	require.Truef(t, ok, "unexpected default mod export type %T", mod.Exports().Default)
 
 	vu.MoveToVUContext()
